First cut of new UI.
